Refactored simple-serializer to smaller files
