+++ lib/serializer/simple-serializer/index.ts
@@ -1,0 +1,1 @@
+export * from './simple-serializer'
